refactor: Update file paths and include directories
The code changes in the Makefile and source files involve updating the file paths and include directories to reflect the new project structure. The changes include renaming the `ServerContainer.cpp` file to `ServerManager.cpp` and updating the corresponding include statement. Additionally, the include directories are modified to include the subdirectories of the `srcs` directory. These changes aim to improve the organization and maintainability of the codebase.

Based on the recent repository commits, the commit message follows the established convention of starting with a verb indicating the type of change (`refactor` in this case) and providing a concise summary of the changes made.
